# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: stories
    description: Story generation operations
paths:
  /stories/generate:
    post:
      operationId: generateStory
      tags: [stories]
      summary: Generate a personalized bedtime story
      description: Uses Claude to write a short bedtime story starring the given child
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateStoryRequest"
      responses:
        "200":
          description: Story generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedStory"
        "400":
          description: Invalid request
        "502":
          description: Generation failed
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    GenerateStoryRequest:
      type: object
      properties:
        childName:
          type: string
        childAge:
          type: integer
        category:
          type: string
        theme:
          type: string
        length:
          type: string
        idea:
          type: string
        interests:
          type: array
          items:
            type: string
      required:
        - childName
        - childAge
        - category
        - theme
        - length
    GeneratedStory:
      type: object
      properties:
        title:
          type: string
        paragraphs:
          type: array
          items:
            type: string
      required:
        - title
        - paragraphs